refactor: separate transport layer from message handling
file rename
