+++ src/tastytrade/common/exceptions.py
