+++ src/tastytrade/dashboard/app.py
@@ -3,7 +3,7 @@ import sys
 
 if __name__ == "__main__":
     from tastytrade.dashboard.dashboard import DashApp
-    from tastytrade.logging import setup_logging
+    from tastytrade.common.logging import setup_logging
 
     # Setup logging
     setup_logging(
