+++ src/tastytrade/messaging/handlers.py
@@ -3,16 +3,16 @@ import logging
 import time
 from dataclasses import dataclass
 from itertools import chain, islice
-from typing import Any, Awaitable, Callable, Dict, Iterator, List, Optional, Protocol, Union, cast
+from typing import Any, Awaitable, Callable, Dict, Iterator, List, Optional, Union, cast
 
-import pandas as pd
-import polars as pl
 from pydantic import ValidationError
 
-from tastytrade.exceptions import MessageProcessingError
-from tastytrade.sessions.configurations import CHANNEL_SPECS
-from tastytrade.sessions.enumerations import Channels
-from tastytrade.sessions.models import BaseEvent, Message
+from tastytrade.common.exceptions import MessageProcessingError
+from tastytrade.config.configurations import CHANNEL_SPECS
+from tastytrade.config.enumerations import Channels
+from tastytrade.messaging.models.events import BaseEvent
+from tastytrade.messaging.models.messages import Message
+from tastytrade.messaging.processors.default import BaseEventProcessor, EventProcessor
 
 logger = logging.getLogger(__name__)
 
@@ -34,58 +34,6 @@ class QueueMetrics:
 
     def record_error(self) -> None:
         self.error_count += 1
-
-
-class EventProcessor(Protocol):
-    """Protocol for event processors"""
-
-    name: str
-    df: pd.DataFrame
-
-    def process_event(self, event: BaseEvent) -> None: ...
-
-
-class BaseEventProcessor:
-    """Base processor that handles DataFrame storage"""
-
-    name = "feed"
-
-    def __init__(self) -> None:
-        self.pl = pl.DataFrame()
-
-    def process_event(self, event: BaseEvent) -> None:
-        self.pl = self.pl.vstack(pl.DataFrame([event]))
-
-        if len(self.pl) > 2 * ROW_LIMIT:
-            self.pl = self.pl.tail(ROW_LIMIT)
-
-    @property
-    def df(self) -> pd.DataFrame:
-        return self.pl.to_pandas()
-
-    def last(self, symbol: str) -> pd.DataFrame:
-        return self.df.loc[self.df["eventSymbol"] == symbol].tail(1)
-
-
-class CandleEventProcessor(BaseEventProcessor):
-
-    def process_event(self, event: BaseEvent) -> None:
-        self.pl = (
-            self.pl.vstack(pl.DataFrame([event]))
-            .unique(subset=["eventSymbol", "time"], keep="last")
-            .sort("time", descending=False)
-        )
-
-    @property
-    def df(self) -> pd.DataFrame:
-        return self.pl.to_pandas().sort_values("time", ascending=True).reset_index(drop=True)
-
-
-class LatestEventProcessor(BaseEventProcessor):
-    name = "feed"
-
-    def process_event(self, event: BaseEvent) -> None:
-        self.pl = self.pl.vstack(pl.DataFrame([event])).unique(subset=["eventSymbol"], keep="last")
 
 
 class EventHandler:
